refactor(#36): 예외처리 로직 코드 exception 패키지로 이동
