+++ src/main/groovy/com/yumyumcoach/domain/challenge/service/ChallengeService.java
@@ -7,8 +7,8 @@ import com.yumyumcoach.domain.challenge.mapper.ChallengeMapper;
 import com.yumyumcoach.domain.challenge.mapper.ChallengeParticipantMapper;
 import com.yumyumcoach.domain.challenge.model.DifficultyCode;
 import com.yumyumcoach.domain.challenge.model.GoalType;
-import com.yumyumcoach.global.error.BusinessException;
-import com.yumyumcoach.global.error.ErrorCode;
+import com.yumyumcoach.global.exception.BusinessException;
+import com.yumyumcoach.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
