+++ src/main/groovy/com/yumyumcoach/domain/community/service/CommentService.java
@@ -8,8 +8,8 @@ import com.yumyumcoach.domain.community.entity.Post;
 import com.yumyumcoach.domain.community.entity.PostComment;
 import com.yumyumcoach.domain.community.mapper.PostCommentMapper;
 import com.yumyumcoach.domain.community.mapper.PostMapper;
-import com.yumyumcoach.global.error.BusinessException;
-import com.yumyumcoach.global.error.ErrorCode;
+import com.yumyumcoach.global.exception.BusinessException;
+import com.yumyumcoach.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
