+++ src/main/groovy/com/yumyumcoach/domain/community/service/PostService.java
@@ -11,8 +11,8 @@ import com.yumyumcoach.domain.community.mapper.PostCommentMapper;
 import com.yumyumcoach.domain.community.mapper.PostImageMapper;
 import com.yumyumcoach.domain.community.mapper.PostLikeMapper;
 import com.yumyumcoach.domain.community.mapper.PostMapper;
-import com.yumyumcoach.global.error.BusinessException;
-import com.yumyumcoach.global.error.ErrorCode;
+import com.yumyumcoach.global.exception.BusinessException;
+import com.yumyumcoach.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
