--- src/main/groovy/com/yumyumcoach/global/exception/InvalidCredentialsException.java
@@ -1,7 +1,0 @@
-package com.yumyumcoach.global.exception;
-
-public class InvalidCredentialsException extends RuntimeException{
-    public InvalidCredentialsException(String message) {
-        super(message);
-    }
-}
